Refactor MultiLibraryHome filtering
